* Поправил тип свойства.

+++ TrafficManagementCenter.Server/Model/TypeAppeal.cs
@@ -5,7 +5,7 @@ namespace Model
     public class TypeAppeal
     {
         [Key]
-        public int Key { get; set; }
+        public long Key { get; set; }
         public string Name { get; set; }
         public string Note { get; set; }
     }
